requirements.txt
python-fasthtml
monsterui
apsw
plotly
pandas
